requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.7.0
numpy==2.1.3
openai==1.53.0
portalocker==2.10.1
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.12.1
setuptools==75.3.0
sniffio==1.3.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
